workers/requirements.txt
# Worker Dependencies
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.4

# Computer Vision & ML
# Using specific compatible versions to avoid dependency conflicts
torch==2.1.0
torchvision==0.16.0
transformers>=4.37.0,<4.50.0
protobuf>=3.20.0  # Required for transformers tokenizers (especially SigLIP)
# clip-by-openai has dependency issues, use direct CLIP from transformers instead
sentence-transformers==2.2.2

# Image Processing
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Audio Processing  
librosa==0.10.1
soundfile==0.12.1
openai-whisper==20240930  # Official OpenAI Whisper package (uses pyproject.toml build)

# OCR (optional - can be installed later if needed)
# paddlepaddle==2.5.2  # Commented out - very large, install separately if needed
# paddleocr==2.7.0     # Commented out - depends on paddlepaddle
easyocr==1.7.0
pytesseract==0.3.10

# Video Processing
ffmpeg-python>=0.2.0

# Scientific Computing
# Pin NumPy to < 2.0 for compatibility with torch, transformers, and other ML libraries
# that were compiled against NumPy 1.x
numpy>=1.24.0,<2.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
aiofiles>=23.2.1
python-dateutil>=2.8.2
structlog>=23.2.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
isort>=5.12.0